+++ test/fixture/define/exports-multiple/output.js
@@ -1,0 +1,3 @@
+var $REF1 = 'baz';
+export var foo = $REF1;
+export var bar = $REF1;