+++ test/fixture/define/exports-overrides/input.js
@@ -1,0 +1,5 @@
+define('foo', ['exports', 'hello', 'world'], function (exports, hello, world) {
+  exports.bar = exports.baz = undefined
+  exports.bar = hello
+  exports.baz = world
+})