+++ test/fixture/define/exports-overrides/output.js
@@ -1,0 +1,4 @@
+import hello from 'hello';
+import world from 'world';
+export var bar = hello;
+export var baz = world;