+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.11.1
+
+- fix: handle multiple exports.foo assignments
+- add: new skipped test case for an edge cases that were not considered before
+
 ## 0.11.0
 
 * edit: return { source, map } instead of { code, map } in the toSource method